fix: Refactor managers into questionset and content packages
