+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/actors/AssessmentItemActor.scala
@@ -12,7 +12,7 @@ import org.sunbird.graph.schema.DefinitionNode
 import org.sunbird.graph.utils.NodeUtil
 import org.sunbird.telemetry.logger.TelemetryManager
 import org.sunbird.utils.RequestUtil
-import org.sunbird.managers.AssessmentManager
+import org.sunbird.managers.questionset.AssessmentManager
 import org.sunbird.validators.AssessmentItemValidator
 import org.sunbird.utils.JavaJsonUtils
 import org.sunbird.utils.AssessmentItemUtils
