+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/v5/managers/AssessmentV5Manager.scala
@@ -10,7 +10,7 @@ import org.sunbird.graph.dac.model.Node
 import org.sunbird.graph.nodes.DataNode
 import org.sunbird.graph.schema.{DefinitionNode, ObjectCategoryDefinition}
 import org.sunbird.graph.utils.NodeUtil
-import org.sunbird.managers.HierarchyManager
+import org.sunbird.managers.questionset.HierarchyManager
 import org.sunbird.telemetry.util.LogTelemetryEventUtil
 import org.sunbird.utils.{AssessmentErrorCodes, RequestUtil}
