+++ assessment-api/assessment-actors/src/test/resources/application.conf
@@ -387,7 +387,7 @@ max.thumbnail.size.pixels=150
 schema.base_path="../../test_schema/"
 
 assessment.skip.validation=false
-uestion.keyspace="dev_question_store"
+question.keyspace="dev_question_store"
 questionset.keyspace="dev_hierarchy_store"
 cassandra {
   lp {
