+++ content-api/content-actors/src/main/scala/org/sunbird/content/util/DiscardManager.scala
@@ -12,7 +12,7 @@ import org.sunbird.graph.OntologyEngineContext
 import org.sunbird.graph.nodes.DataNode
 import org.sunbird.graph.dac.model.Node
 import org.sunbird.graph.external.ExternalPropsManager
-import org.sunbird.managers.UpdateHierarchyManager.{fetchHierarchy, shouldImageBeDeleted}
+import org.sunbird.managers.content.UpdateHierarchyManager.{fetchHierarchy, shouldImageBeDeleted}
 import org.sunbird.telemetry.logger.TelemetryManager
 import org.sunbird.utils.{HierarchyConstants, HierarchyErrorCodes}
