+++ content-api/hierarchy-manager/src/test/scala/org/sunbird/managers/TestUpdateHierarchy.scala
@@ -1,4 +1,4 @@
-package org.sunbird.managers
+package org.sunbird.managers.content
 
 import java.util
